+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/area/SpawnType.kt
@@ -1,0 +1,5 @@
+package com.github.morinoparty.morinoinsect.catching.area
+
+enum class SpawnType {
+    ONGROUND, ONTREE
+}
