+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/condition/TimeCondition.kt
@@ -1,5 +1,6 @@
 package com.github.morinoparty.morinoinsect.catching.condition
 
+import com.github.morinoparty.morinoinsect.catching.area.SpawnType
 import org.bukkit.Material
 import org.bukkit.entity.Player
 
@@ -9,7 +10,7 @@ class TimeCondition(
     override fun check(
         catcher: Player,
         block: Material,
-        spawnType: String
+        spawnType: SpawnType
     ): Boolean {
         return catcher.world.isDayTime == state
     }
