+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/insect/InsectCondition.kt
@@ -1,7 +1,9 @@
 package com.github.morinoparty.morinoinsect.catching.insect
 
+import com.github.morinoparty.morinoinsect.catching.area.SpawnType
 import com.github.morinoparty.morinoinsect.catching.condition.BlockCondition
 import com.github.morinoparty.morinoinsect.catching.condition.Condition
+import com.github.morinoparty.morinoinsect.catching.condition.SpawnTypeCondition
 import com.github.morinoparty.morinoinsect.catching.condition.TimeCondition
 import kotlinx.serialization.Serializable
 import org.bukkit.Material
@@ -27,7 +29,8 @@ data class InsectCondition(
     fun generateConditionSet(): Set<Condition> {
         return setOfNotNull(
             BlockCondition(blocks.map { Material.valueOf(it.toUpperCase()) }),
-            TimeCondition(time?.replace("day", "true")?.replace("night", "false").toBoolean())
+            TimeCondition(time?.replace("day", "true")?.replace("night", "false").toBoolean()),
+            SpawnTypeCondition(SpawnType.valueOf(spawnType.toUpperCase()))
         )
     }
 }
