--- src/main/kotlin/com/github/morinoparty/morinoinsect/insectslist/InsectTier.kt
@@ -1,6 +1,0 @@
-package com.github.morinoparty.morinoinsect.insectslist
-
-enum class InsectTier {
-    // ToDo: Tierはコンフィグから追加可能な変数にしておく　＜ー　今namiuさんがやってくれてる
-    NORMAL
-}
