--- src/main/kotlin/com/github/morinoparty/morinoinsect/insectslist/SpawnType.kt
@@ -1,6 +1,0 @@
-package com.github.morinoparty.morinoinsect.insectslist
-
-enum class SpawnType {
-    ONGROUND,
-    ONTREE
-}
