--- src/main/kotlin/com/github/morinoparty/morinoinsect/util/Vector.kt
@@ -1,8 +1,0 @@
-package com.github.morinoparty.morinoinsect.util
-
-// 範囲指定の時に必要なので作った
-data class Vector(
-    val x: Int,
-    val y: Int,
-    val z: Int
-)
